DAOS-7238 test: check pool create svc_reps length for Coverity issue
As part of dmg_pool_create() flow, following successful execution
of the dmg pool create command, it is expected that the svc_reps
returned by dmg shall be of length > 0. Before this change, that is not
checked in the code, and Coverity CID 313058 suggests that
a use after free could occur with outpool->svc in the daos_test.
This change adds a check in parse_pool_info, returning -DER_INVAL
when the length of the svc_reps json array is less than or equal to 0.
If this were to actually occur, the dmg_pool_create() function would
return, instead of calling the d_rank_list_copy() that the CID suggests
would cause the list to be freed (as part of a realloc).

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>