DAOS-6155 swim: Allow the DEAD to resurrect
We use SWIM only to monitor group members. When SWIM confirms the death
of a member, the MS decides if, when, and in what group version to
remove the member from the group, so that the group membership is
strongly consistent and suitable for spanning-tree-based operations. If
auto exclusion is off, or if the MS decides not to exclude too many
engines, then DEAD members may remain in the group. If such members are
not actually dead, they need to learn that others consider them DEAD.
And, they need to be able to clarify that they are now ALIVE again.

This patch enables DEAD members to continue swimming. Since DEAD updates
may have been dropped, swim_updates_send will make sure that the
recipient's status is always sent. When a member learns that it is
considered DEAD, it will bump its incarnation. DEAD status can already
be overridden by other status with higher incarnations.

Note that a required change not included in this patch is that the self
incarnation must start from a higher number than previous ones, instead
of 0. This change will be made by a later patch.

Signed-off-by: Li Wei <wei.g.li@intel.com>