DAOS-7286 duns: resolve path where the entry can be anywhere in the path
Today duns_resolve_path() works only when the duns entry is the last
one in the path provided, and fails otherwise. This PR would keep
doing a reverse lookup till it finds a duns entry and then fails
if it doesn't.

Also fix an issue in daos cont query to check the container type using
the property rather than the UNS entry to avoid segfault when a direcy
path is specified.

Features: datamover,-hw
Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>