Cart-963 CartTest: Re-order build sequence.
The order of building product was put back the way it was prior to
the start of PR.  Originally, cart_utils library was build as part of
cart test, so building of utilities which was now dependent on the
library had to come afterwards.  cart_utils library is now built as
cart, so the build of utilities can be place back where it was.  This
is an attempt to fix dependency issues when building CentOS RPMs.

Signed-off-by: Steven Bollinger <stevenx.bollinger@intel.com>