DAOS-6799 build: Enable separate builds for DAOS components
Right now, there is just client, server, and test but we can
potentially extend this to a larger set of targeted builds.

It adds some dummy targets (server, client, and test) which
are checked at runtime to limit the set of things we add
to the build.

For example
scons client -j24
scons client install -j24

Will just build and install client components

This is basically a functioning proof of concept

Fixes DAOS-6864.  This was caused by linking libdaos.so into
the daos_engine.  This library depends on libdaos_common.so
but the engine already links the pmem enabled equivalent.

Rather than link libdaos.so from the engine, I've modified
the build to link in the relevant files instead so there
is no longer a dependence on libdaos.so from daos_engine.

TODO: We may be able to clean up the rpms a bit as, in theory,
libdaos_common.so should be a client only library.

Fix a few pylint issues

An xml named directory caused an exception in fix_cmocka_xml.py

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>