DAOS-623 ci: Update old test_tag logic
The pipeline library has been updated to properly set TEST_TAG for any
of the Test-tag-* pragmas, so it doesn't need to be done in
test_main.sh any more.

Require that the caller set TEST_TAG and error out if they don't.  There
is no reasonable default for this value.

Skip-PR-comments: true
Test-tag-hw-small: pr datamover

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>