DAOS-7128 test: Improve NLT/junit handling.
Keep a valid junit file with a failure saved at all times,
then on successful exit remove the failure and re-write
the file.  Previously this was done the other way round,
however on failure the code is called during shutdown
and junit-xml is unable to write the file.

Add some debugging.

Parallel-build: true

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>