DAOS-7489 tests: Add launch.py option to repeat test
Adding the '--repeat <int>' option to the launch.py command to allow the
user to repeat the execution of the tests <int> times.  The default is
to run the tests once.  This command line option can be used in CI
through the 'Test-repeat: <int>' or 'Test-repeat-<stage>: <int>' pragma.

Test-tag: test_pool_svc iorsmall
Test-repeat-vm: 3

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>