DAOS-623 actions: Improve github actions support.
Add another build test to github-actions.  This doesn't
modify the existing code, but intends to replace it once
stable.

Use the github cache to re-use docker images for speed.
Do a single image create, then re-use this for different
compilers etc.
Run NLT in docker, and report server leaks as annotations.
Remove PSM2 from the Docker builds.

This is enabled to build centos.8 with gcc/clang, and to
build and test with ubuntu/clang but can be readily
extended to the rest of the matrix once landed and stable
without adding additional time or complexity.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>