DAOS-623 test: Improve test stability in GA.
Increase shutdown timeout in dmg from 10 to 25 seconds to
avoid false failures in Github Actions.  This timeout is
scheduled for removal soon, but in the restrictive
environment of Azure the current value is not long ehough
and we're seeing occasional failures.

Run test in a loop to see how frequently we are hitting
this timeout.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>