DAOS-6864 build: Don't use libdaos.so from server (#5706)
libdaos.so depends on a conflicting version of libdaos_common
with pmem enabled.  We don't want two versions of
libdaos_common pulled in at the same time.

Instead, statically link all of the libdaos targets into
the daos_engine binary and use libdaos_common_pmem.so
to resolve the needed symbols.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>