DAOS-7174 container: open by label support
With this change, libdaos clients may open a container by specifying
only a label. This is an alternative to opening by UUID. A new API
daos_cont_open_bylabel() is provided that uses a new CaRT interface
CONT_OPEN_BYLABEL.

With this change, the libdaos minor API version is updated from
1.2.1 to 1.2.2.

Also with this change, the daos_engine pool and container service
metadata versions are incremented from 1 to 2. This is to reflect
the addition of a new container UUIDs KVS that maps label to UUID.
Container creates with non-default label properties included
in the request result in insertions to the new KVS, and
CONT_OPEN_BYLABEL operations perform lookups in this KVS, before
the "by-UUID" lookup in the container KVS used in all cases.

Also with this change, the daos_test container properties test creates
a container with a label property, and exercises daos_cont_open_bylabel().

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>