DAOS-7480 pool: add co_in_ver to pool map
Add co_in_ver to the pool map, which will share with co_out_ver
to indicate when the target is being reintegrated or extended,
so if the leader is switched or restarted, it can retrieve the
extend or reintegrate job version correctly.

Add more debug information to the rebuild process.

Signed-off-by: Di Wang <di.wang@intel.com>