serialize: create daos_serialize library
The daos_serialize library is generated if hdf5 is found
on the system, and if it is not then no library is
generated, and it is added as an optional sub package to
the DAOS spec file.

The library will be used to preserve daos
container properties and user attributes using the
daos filesystem tool when data is moved from DAOS to a
regular POSIX filesystem. The library will also be used
to store serialization and deserialization functions for
the future daos serialize and daos deserialize utilities.

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>