DAOS-7666 dtx: properly cleanup DTX sub requests
For efficiently sending DTX RPCs to related servers, we classify
the DTX sub requests in DRAM on the leader. For each sub request,
usually, it will be added in the classification tree and a global
list. But under some special cases, such as out of memory on the
leader, above classification may fail, as to related sub request
only exists in the list. Then only destroying the classification
tree cannot release related sub request. We still need to travel
the global list to drop those failed sub requests after DTX RPC.

Signed-off-by: Fan Yong <fan.yong@intel.com>