DAOS-7492 dfuse: Do not set max_read or max_write.
Rather than setting a desired value for max_read or max_write
just go with the defaults.

On newer systems this will default to the largest possible size
for read, on older systems this will default to 4k.

This avoids a issue on CentOS 8 where dfuse was requesting
a larger size for perforamnce, however CentOS 8 has a newer
kernel which checks buffer size and a older libfuse which
doesn't, so requesting a larger size on CentOS 8 results
in the kernel rejecting fuse requests.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>