DAOS-7653 test: Update NLT to not check for dmg return code.
A recent change to dmg meant that some commands return with
an exit code of 0 indicating success.  NLT was incorrectly
returning in this case, so was not retrying format commands.

Remove the incorrect check from NLT, and improve timing of
server startup to make it quicker by not waiting so long
between issueing commands.

Parallel-build: true

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>