DAOS-7174 container: new open by label API (#5784)
With this change, libdaos clients may open a container by specifying
only a label. This is an alternative to opening by UUID. A new API
daos_cont_open_bylabel() is provided that uses a new CaRT interface
CONT_OPEN_BYLABEL. The libdaos minor API version is updated to 1.2.2.
The daos_test container properties test creates a container with
a label property, and exercises daos_cont_open_bylabel().

Also with this change, the daos_engine pool and container service
metadata versions are incremented from 1 to 2. This is to reflect
the addition of a new container UUIDs KVS that maps label to UUID.
Container creates with non-default label properties included
in the request result in insertions to the new KVS, and
CONT_OPEN_BYLABEL operations perform lookups in this KVS, before
the "by-UUID" lookup in the container KVS used in all cases.

Also with this change, both for pool connect bylabel and this
contanier addition, the task interfaces are simplified. The existing
daos_pool_connect_t and daos_cont_open_t structures are used,
adding a label at the end of those structures for ABI compatibility
with existing callers. i.e., there are no new structures and task
table entries (i.e., daos_pool_connect_lbl_t is removed in this patch).

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>