DAOS-7754 test: reduce large pool SCM size in basic_snapshot
Before this change, in virtual machine (VM)-based testing of the
container basic snapshot tests, those tests with the "8gb" scm_size
were failing due to the Linux oom-killer terminating the daos_engine
process. Currently 12GB of memory is provisoned for the VMs, and recent
changes to the engine implementation consume more memory, likely
putting too much pressure on the VM.

With this change, the "8gb" SCM size test cases are changed to instead
allocate 7GiB of SCM for the pool.

Test-tag: pr daily_regression

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>