DAOS-7174 pool,container: increment md layout low versions (#5814)
Update the minimum-supported pool and container metadata layout
version numbers to version 2. This corresponds to the new
container metadata layout that adds a container label -> UUID KVS.

With this change, pools or containers created in earlier versions
of DAOS (e.g., DAOS-1.2 and earlier) will not be usable by newer
versions of DAOS software. Such backward compatibility may be added
in a future version.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>