DAOS-7174 dfuse: Allow dfuse to accept pool labels on command line.
Update dfuse so that pool labels can be passed on the command
line instead of uuids.

Update NLT to overhaul pool handling code.  The idea here is
to have it create a pool with a label once, and re-use that
pool over multiple invocations, however DAOS-7805 makes that
difficult so for now create a pool with the label when required
and assume rather than query the label.

This does not allow labels to be used as intermediate paths
which running dfuse without a pool, the uuid is still
required for that.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>