DAOS-7489 tests: Add launch.py option to repeat test (#5678)
Adding the '--repeat <int>' option to the launch.py command to allow the
user to repeat the execution of the tests <int> times.  The default is
to run the tests once.  This command line option can be used in CI
through the 'Test-repeat: <int>' or 'Test-repeat-<stage>: <int>' pragma.

Ensuring the renaming of the avocado test directory for Jenkins results
in unique directories for repeated tests by including the repeat number in
the renamed directory.

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>