DAOS-7819 daos: add dst create tests for fs copy tool
DAOS-7455 ticket fixed the destination container create
in fs copy for cont -> new cont but missed fixing the case
for when the source is a POSIX directory (not a DAOS conainter).

This commit also adds functional tests for fs copy for dm_dst_create to verify
all dst create cases are working for fs copy.

Test-tag: pr dm_dst_create fs_copy

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>