DAOS-7821 ga: Improve github actions coverage and resiliance.
Use ubuntu 21.04 in GA for testing as it's newer.
Remove psm2 and psm3 from the build.
Seperate codespell into it's own workflow.
Update the newer process to build Java in docker.
Remove the CentOS 8 and Ubuntu 20.04 builds from older
process as it's now replicated in the newer one.
Report high priority error messsages to github as annotations.
Add missing packages on CentOS 8

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>