DAOS-7869 dfuse: Add a --path option to dfuse and clean up arg parsing.
Add the option to pass in a path from which UNS should be
queried but which isn't the mountpoint.  Keep the old behaviour
of also loading this from the mountpoint itself.

This allows:
dfuse --path /mount/dir-a --mountpoint /mnt/dir-b

Do not allow pool and container information to be set multiple
ways, either command line, the path or the mountpoint.

Move some variables from struct entries to locals.
Resolve coverity issue 332230

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>