DAOS-7936 vos: Delete records are not punches
The prior patch that modified the implementation of vos_obj_array_remove
will break snapshots because entries are treated as an ordinary punch
whereas they are really intended to delete an epoch range.

Modify the function and algorithm accordingly.

1. Change evtree iterators so that as new flags are added, old ones
are implied.   For example, there is no way to view only covered
entries.  This is just a simplification that allows me to always
sort the delete records at the end.
2. As such, aggregation can simply request VOS_ITER_RECX_DELETED
and it will get COVERED and VISIBLE entries
3. Add a pass to the sort algrorithm to "cover" entries that are
marked as deleted by a delete record.
4. Modify vos_obj_array_remove to delete full extents and add this
special punch record for partially covered extents in the epoch range
5. Modify evt_insert to allow modification of delete record

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>