DAOS-7894 agent: Refactor GetAttachInfo cache
Split the agent cache into two separate caches: one for the remote
GetAttachInfo response, and one for the fabric interface topology
on the machine. The remote cache may still be disabled using the
environment variable. The local fabric cache cannot be disabled,
as the fabric topology is assumed to be consistent during the life
of the agent process.

Signed-off-by: Kris Jacque <kristin.jacque@intel.com>