DAOS-7174 pool,container: add common label verify function (#6007)
With this change a public function daos_label_is_valid() is added
to check whether a label string conforms to a maximum length
requirement, and a matches a specific pattern of legal characters
(defined using a regex). The daos_test pool tests are expanded to
exercise the function with a few valid and invalid labels.

The default label (not set) property strings for pools and containers
are modified so they can be matched by the regex. This patch also
changes the engine pool and container create handlers to verify the
user-provided properties, not just the default properties duplicated
in the initial step.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>