DAOS-7420 vos: Add object specific mode to vos_discard
Also, fix vos_discard so that it leaves subsequent commits
visible.   This is done by scanning to DAOS_EPOCH_MAX
and then tracking whether the subsequent updates may
have bypassed adding an ilog entry due to an existing
creation entry.   If that entry is discarded, we need
to update the ilog accordingly to ensure we don't have
orphaned entries.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>