DAOS-7891 daos: Update expected retrun from dfs_cont_create()
Change the server to reply -DER_EXISTS on attempt to create a
container where the label already exists.
Handle the return code from dfs_cont_create() as a system
error number.
Handle container exists in NLT by checking the error, deleting
the container and re-trying the create.
Turn off valgrind for some invocations of daos cont
create/destroy in NLT to improve run-time.

This code will only activate if NLT is run on a system
with pre-existing containers so it does not get executed in
CI, and currently fails as container destoy by label does
not yet work.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>