DAOS-6540 dfuse: Detect clobbered files and update kernel accordingly.
Use the oid option to dfs when calling move or rename to detect
what if any file has been removed, and check for that file in
the local inode table.
When files have been moved or are observed at a different location
to what dfuse expects then update the kernel to remove the old
dentries which should reduce kernel resources required.
When files are removed then check if the file is known and if so
mark is as removed.
For getattr on removed files return most recent data rather than error.
For setattr on removed files allow setting of times only as this can
happen with cached writes.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>