DAOS-7174 pool: take a generic string in pool_connect
Change daos_pool_connect() to take a string an a input parameter
that can be interpreted as either a label or a UUID string.

Backward API compatibility is supported via a macro that will
automatically convert a uuid_t (old API) to a UUID string.
We could eventually remove the macro in 2.2 or beyond.

Backward ABI compatibility is supported by keeping around
the daos_pool_connect symbol, but not advertising it via the
header files (the macro overrids it). We could also remove this
symbol at some point when all users have recompiled against 2.x
code.

Handle backward compatibility for the task API too.

Signed-off-by: Johann Lombardi <johann.lombardi@intel.com>