DAOS-8034 il: Do not attempt to use interception of special files.
If the filename starts with /proc, /sys or /dev then do not try the
ioctl to check for dfuse.  This should cut down on I/O calls
which were destined to fail and make program startup in particular
a little bit faster.
Make the logging for non-intercepted files more consistent across
open,fopen and creat.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>