DAOS-623 test: Add better exception handling to NLT.
Run all of NLT under an exception handler, and if there is an
exception record it as a failed test case which includes both
the stack trace and also the exception string.

This helps a lot with unexpected failures in the NLT code, at
least giving a stack trace, and depeending on the exception
a text description of the problem as well.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>