DAOS-7951 daos_test: Add VMD testing to daos_test NVMe recovery tests
New C APIs to allow for testing VMD LED functionality:

- int daos_mgmt_get_led_state(const char *group, uuid_t devuuid, int *led_state, daos_event_t *ev)
Will return the current LED state of the VMD status LED on the device.

- int daos_mgmt_reset_led_state(const char *group, uuid_t devuuid, daos_event_t *ev)
Will reset the LED state of the VMD device back to off.

Updated current tests in daos_nvme_recovery.c to test for VMD functionality:
- dmg_storage_device_list() now returns the traddr addressed from the NVMe devices
- NVMe Recovery 4: Verify device states via log after NVMe SSD eviction: now queries
for the LED state of the VMD device after it has been evicted, and then resets it
back to off. If the device is not a VMD device, LED testing is skipped.
- NVMe Recovery 5: Verify device states after NVMe SSD eviction/reint: now adds an
evicted device back into use by replacing the device with itself (excluding target
reintegration). Device states are verified after reintegration. LED states are also
verified before and after reintegration if the evicted device is a VMD device.
- NVMe Recovery 7: Identify an SSD via LED (VMD only): new test added, sets the
IDENTIFY state of the status LED on all VMD devices present (quick blink state).
Test checks that LEDs are returned to off after the 60 sec duration. Skips if
no VMD devices are present.

Test-tag: pr daos_core_test_nvme

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>