DAOS-7821 gha: Add killing of previous jobs to gha.
Update github actions workflows to use the new concurrency
feature, and to have them kill old jobs for the same
reference that are still running.
This should reduce the load that we run in github actions
and go some way to reducing the turn-around time, even
more so when the queues are long and it becomes more likely
that there is a new push to a PR before the previous build
has completed.
Do not run the workflows on landings to master, this will
again reduce the load on GHA and result in faster time to
result, but also there does not seem to be a way to apply
the concurrency/kill-previous-runs feature selectively to
PRs only and not branches.

Remove some more unused packages from the dockerfiles,
and update the pip command to be more correct.
Remove the spdk re-build now that DAOS-7764 seems fixed.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>