New API for pipeline executions in DAOS
This is an API proposal to support the execution of data
pipelines in DAOS. Comments, suggestions, corrections, etc. are
all welcome.

The idea behind this proposal is for DAOS to support some data
filtering/aggregation operations in order to reduce data transfer
(from DAOS to the client). The first intended application for
this is MariaDB.

Skip-test: true

Signed-off-by: Eduardo Berrocal <eduberrocal@gmail.com>