DAOS-7174 cont: take uuid as output param in cont_create
Change daos_cont_create() to take a uuid as output parameter.
The UUID cannot be provided by the user any longer and will
be returned to the caller.

Backward API compatibility is supported via a macro that use
the old behavior when a uuid_t is passed instead of a uuid_t *.

Backward ABI compatibility is supported by keeping around
the daos_cont_create symbols, but not advertising them via
the header files (the macro overrides it).

Rename daos_cont_create_by_label() to daos_cont_create_with_label()
to align with new function naming.

Signed-off-by: Johann Lombardi <johann.lombardi@intel.com>