attempt to fix groupupdate race
a rare race where async groupupdate requests come in while
join requests are being services could result in an attempt
to perform a group update with the same map version, which
triggers an assertion in the engine. this fix records the
last map version used to successfully perform a group update
and ignores duplicate attempts.

Quick-functional: true
Test-tag: ms_resilience
Test-Repeat: 40

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>