DAOS-623 test: Add --failfast flag to launch.py
If set, appends '--failfast on' to the avocado run
arguments. This will cause avocado to abort the test
run at the first test failure. Also breaks out of the
repeat loop if --repeat was specified.

Fixes a corner case in the system query helper where the
response succeeds but contains a null members array.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>