DAOS-8210 container: return -DER_NOMEM
The return code is not being set when out of memory error occurs. This was
revealed in fault injection testing for filesystem copy which calls
dfs_cont_create, which then calls daos_cont_create...which eventually calls this
function and returns the incorrect error code. This is causing the filesystem
copy fault injection tests to report DER_INVAL value when it should be returning
DER_NOMEM.

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>