DAOS-8173 test: Use positional parameter for daos container command and fixed tests
Use CommandWithPositionalParameters in daos command.
Use PositionalParameter for self.pool since most of the daos container
subcommands have pool as positional parameter.
Reordered the parameters and the methods to match that of the help.

Renamed pool_list_cont to container_list and use JSON. Fixed the tests to
incorporate this change.

Skip-unit-test: true
Test-tag: pr dmg_pool_evict list_containers
Signed-off-by: Makito Kano <makito.kano@intel.com>