DAOS-7987 test: fix CID 329541, 331438, 329382
329541 In cmd_line_parse: The string buffer may
not have a null terminator if the source string's
length is equal to the buffer size

Set source buffer to max - 1 to make so there is room
in destination buffer for null byte.

331438 In server_main: A character buffer that has
not been null terminated is passed to a function
expecting a null terminated string

Set null terminator on string.

329382 In parse_verify_swim_status_arg: Reads an
uninitialized pointer or its target

Initialize the array.

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>