DAOS-6726 test: Update Jenkinsfile to fail build on memory leaks.
Now that we have resolved all detected server and client memory
leaks change Jenkinsfile so that any leaks cause a build failure.

Until now we've had either leaks or intermittent leaks so this
check has been on but not enforced which has lead to a number
of regressions slipping in.  As master is now routinely showing
no leaks discovered them change the settings to enforce this.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>