DAOS-8257 test: Remove pool/connect_test.py because it's invalid (#6420)
It's an old test and several changes to DAOS and the test code in
the past two years made it invalid.

It's testing the following:
1. Pool create with -s.
Maybe it's a system name. It's from the old daosctl command.
daosctl create-pool -s "daos_server/rubbish"

2. Check that the pool directory is created in the mount point.

3. Connect with the test system name.
daosctl connect-pool -s "daos_server/rubbish"

For 1, I there's already a test that does the negative name testing in
pool/destroy_tests.py test_destroy_invalid_group(). I think we used to
be able to pass in the name to daosctl during create, but now we need
to define it in daos_control.yml instead of passing it into dmg.

For 2, there are a few tests that test this.

For 3, now we have to use API for connect, but there doesn't seem to be
a parameter for name. (Note, the original code was only doing positive
testing. i.e., not creating with daos_server and connecting with rubbish.)

Thus, this test is invalid and should be removed.

Signed-off-by: Makito Kano <makito.kano@intel.com>