DAOS-8259 rebuild: ignore DER_EXIST for reintegration
If the reintegrate is restarted due to the failure, then reclaim
phase will not be executed, so it might put multiple shards in
the same VOS target, so it might see duplicate OID in the next
scan phase, i.e. see -DER_EXIST failure, so let's ignore this
failure for the moment, since the duplicate object will be removed
in the final reclaim phase anyway.

Signed-off-by: Di Wang <di.wang@intel.com>