test: update dm_large_file and dm_large_posix yaml files
The dm_large_file test is intermittently failing with a DER_NOSPACE
error when the pool is being created. This updates the yaml to
instead use percentages for scm and nvme so that the test will
not try to create a pool that is too big for the CI cluster.

The dm_obj_large_posix test is intermittently failing with a
timeout. The cont_clone portion alone was taking about 5
minutes on successful runs, and the timeout limit only had
about 21 seconds left. This updates the yaml to give the
cont_clone test two more minutes to complete.

Test-tag: pr datamover

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>