DAOS-8298 dfuse: Return a valid duns xattr for all container roots.
Rather than have container create sometimes set an xattr on the root
directory when a container is created do not create this xattr, but
rather have dfuse create the xattr on the fly when one is requested.
Do not allow the xattr to be set on container roots.

This allows commands like "daos fs get-attr" to work correctly,
regardless of how the container was created.

Add a test for this, and correct some returncodes from the "fs get-attr"
error paths.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>