DAOS-8332 vos: Size of allocation needs to be based on real merge window
In prior patch, the merge window size was adjusted based on the newly
found visible logical extent.   The buffer size allocated needs to
be based on the actual size of enclosed extents.   Track the allocation
hi value for this purpose.

This bug was exposed in EC aggregation.  Add some testing for
aggregation with parity ranges and fix a couple of remaining issues
this change exposed.   It's still possible to get a new logical extent
that has already been removed so modify the assertion accordingly
to check if it's covered by a removal record.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>