Improve the ENOSPACE test.
Verify that the file size is correct after both successful and
unsucessful writes.

Improve error handing using try/finally blocks so that objects
can be closed in the right order with any warnings being
propagated correctly.

The new test is failing with a file size issue, and even if it
wasn't it triggers a number of server warnings.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>