DAOS-8313 control: Minimize scope of cleanup on pool create
When a pool is found to already be in the Creating state
on a retried create, trigger the cleanup code for that single
pool.

Also adds -DER_UNREACH to the set of retryable pool errors,
as testing shows that these can still be encountered after
a server has been killed but hasn't been excluded yet.
Removes some redundant retry code from the test harness.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>