DAOS-7479 container: query optionally retreive snapshots
For daos_cont_query(), if daos_cont_info_t.ci_snapshots is not NULL
fetch and return to the caller the number of, and list of snapshot
epochs created for the container.

Container CaRT RPC version update from 4 to 5 to reflect the addition
of a bulk handle (input) and snapshot count (output) in CONT_QUERY.

API version update to v1.5.0 with addition of daos_cont_info_t struct
ci_snapshot_names member.

Go-based daos utility code changes to allocate ci_snapshots if invoked
with the --verbose command-line option. And to retry the CONT_QUERY
RPC in the event that the snapshots storage is insufficient to store
the full list of container snapshots.

daos_test snapshot test additions to exercise query for snapshots,
alongside the existing list_snaps API.

Future changes will be required
- container open API similarly allows retrieval of snapshots
- for query, open, and list_snapshot, retrieve snapshot names
- engine container metadata format (+ version update): snapshot names

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>