Code drop for pipeline API implementation
This is the first code drop for the implementation
of the pipeline API. This code still does everthing
in the client side and still does not implement all
functionality (like filtering by Object ID).

The goal of this first iteration is to make sure the API
does what is supposed to do, without performance considerations.
It does everything on the client side using the daos_obj API.

Skip-test: true