DAOS-8406 vos: Check evtree for orphaned data
We can have orphaned removal records even if there is no
orphaned data.   Augment the check for an empty tree
to only fail if there there is actual data and the tree
isn't just removal entries.  This case can happen if
the EC aggregation inserts removals while aggregation
is processessing the tree.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>