DAOS-8425 vos: Check evtree for orphaned data (#6573)
* We can have orphaned removal records even if there is no
orphaned data.   Augment the check for an empty tree
to only fail if there there is actual data and the tree
isn't just removal entries.  This case can happen if
the EC aggregation inserts removals while aggregation
is processessing the tree.
* Add some logs so we can see if this case is happening and
debug it if it's unexpected.   There may be legit cases where
it happens so don't treat it as a bug until we can prove
otherwise.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>