DAOS-8356 test: update datamover tests to reflect uns changes
    * comment out datamover tests that use dfuse as an emulated
      shared POSIX filesystem for dm_large_file/dir tests until
      there is a shared POSIX filesystem to run these tests across
      nodes in CI.
    * update dm_negative.py and dm_posix_meta_entry.py to use
      the local POSIX filesystem instead of using dfuse

These updates are necessary because duns_resolve_path will now
resolve paths in a dfuse mountpoint that it did not before.
Previously, a dfuse mounted "/tmp/dfuse" would not be resolved
to a DAOS path, but now it is, so datamover tests now need to
use real POSIX paths.

Test-tag: datamover
Quick-Functional: true

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>