DAOS-8460 vmd: Add root privileges to detectVMD() and remove device ID
$lspci command does not work under normal user in SUSE. Currently VMD
devices are queried via lspci, and on OpenSUSE no VMD devices are
returned as the command fails. Add sudo to lspci command, and remove
VMD PCI device ID (0x201d) from query as the ID is specific to SKX
platforms. Instead only query for "Volume Managament Device", as that
remains the same in ICX systems when parsing for VMD addresses.

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>