DAOS-8449 object: object may be removed during DTX sync
There is CPU yield during DTX resync. Then it is normal that
someone may has removed/punched the object during the interval
of resync the DTX enteries that are related with the object and
marking the object as re-synced.

On the other hand, try to verify an object that does not exists
is not fatal failure.

Test-tag-hw-large: pr,hw,large nvme_pool_exclude

Signed-off-by: Fan Yong <fan.yong@intel.com>