DAOS-8432 Remove R: hwloc
It should never have been specified as an explicit Requires: as RPM's
autoprovides/requires will take care of installing the necessary hwloc
library RPM at installation time.

Also Remove hwloc from the required packages for functional testing.

This is all because we are going to ship compat-hwloc1 with DAOS to be
able to use the RPMs that we build on EL 8.3 on EL 8.4 also.

Skip-PR-comments: true
Skip-func-test-vm-valgrind: false
Skip-func-test-el8: false
Skip-func-test-leap15: false

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>