DAOS-7424 engine: Reduce drpc_client_tests delay
When run under Valgrind, test_drpc_verify_notify_pool_svc_update
sometimes takes 10 s or more to return from dss_drpc_call. It is unclear
what exactly causes this delay, but adding a system call into the
polling loop in dss_drpc_call seems to help. (Probably Valgrind has some
impact on memory model and/or scheduling.) This patch adds a real thread
sleep to the sched_req_sleep mock (which is not supposed to block the
current thread) as "the system call".

Signed-off-by: Li Wei <wei.g.li@intel.com>