DAOS-8497 test: Return default values for file and line
In prior patch #6627, if truncation causes bad formatting on a
line, we can be defensive and just return 'Unknown' and 0 for
file and line number rather than raising an exception.

Ideally, we wouldn't hit this problem at all but in instances
where servers are killed and/or restarted, it can happen and
isn't really an error.

Test-tag: offline_reintegration_full

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>