DAOS-8437 pool: fix rdb ref leak in pool extend (#6611)
Before this change, test osa/osa_dmg_negative_test.py
(tag dmg_negative_test_extend) when issuing dmg pool extend commands
that refer to existing pool ranks (e.g., 1, 2, or 3), goes through
a code path in the pool service that leaks a reference on struct rdb.
The pool_extend_internal() function calls rdb_tx_begin() incrementing
the reference, but the pool_extend_map() function it calls invokes
rdb_tx_end() that would decrement the reference. In this scenario,
the rdb_tx_end() call is missed in the error handling path. At the end
of the test pool destroy fails since the RSVC_STOP broadcast does not
get a reply from the pool service leader (it is waiting on references
that will never be released).

This change promotes the rdb_tx_end() call up to pool_extend_internal()
so that the db reference is released in successful and error scenarios.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>